+++ src/onegov.town/onegov/town/app.py
@@ -7,8 +7,6 @@ use different templating languages.
 
 """
 
-import transaction
-
 from cached_property import cached_property
 from collections import defaultdict
 from contextlib import contextmanager
@@ -19,8 +17,6 @@ from onegov.page import PageCollection
 from onegov.search import ElasticsearchApp
 from onegov.shared import asset
 from onegov.ticket import TicketCollection
-from onegov.town import log
-from onegov.town.initial_content import add_builtin_forms
 from onegov.town.models import Town, Topic
 from onegov.town.theme import TownTheme
 from webassets import Bundle
@@ -169,16 +165,6 @@ class TownApp(Framework, LibresIntegration, ElasticsearchApp):
                 s for s in self.session_manager.list_schemas()
                 if s.startswith(schema_prefix)
             )
-
-            for schema in self.known_schemas:
-                self.session_manager.set_current_schema(schema)
-
-                session = self.session()
-                add_builtin_forms(session)
-                session.flush()
-                transaction.commit()
-
-            log.info('Updated all builtin forms')
 
     @property
     def theme_options(self):